requirements.txt
flask
mysql-connector-python
sentence-transformers
chromadb
python-dotenv
pytest
numpy
transformers